# [WIP][SPARK-31919][SQL] Push down more predicates through Join