@maropu 2020-06-06
Just to check; it seems this PR doesn't depend on CNF conversion, but we could get the totally same performance gains with https://github.com/apache/spark/pull/28733?